Add : data-h2.sql의 total_score + userquizhistoryservice에서 solvedScore&total_score

+++ src/main/java/com/example/demo/entity/UserEntity.java
@@ -37,7 +37,7 @@ public class UserEntity {
 //    private String activation;
     @Column(name = "activated")
     private boolean activated;
-    private Integer totalScore;
+    private float totalScore;
     @Column(nullable = false, updatable = false, insertable = false, columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
     private Timestamp registerDate;
